requirements.txt
appnope==0.1.0
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
cryptography==1.9
decorator==4.0.11
Django==1.8.18
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-compressor==2.1.1
django-coverage-plugin==1.5.0
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
django-grappelli==2.10.1
django-haystack==2.6.1
-e git://github.com/MongoEngine/django-mongoengine.git@fcfce1e351a1fc0f8723346871e3c7b211f1c943#egg=django_mongoengine
django-mptt==0.8.7
django-nose==1.4.4
django-pandas==0.4.3
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
docopt==0.6.2
Fabric==1.13.2
factory-boy==2.8.1
Faker==0.7.15
fire==0.1.1
idna==2.5
ipython==5.4.1
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
mongoengine==0.13.0
nose==1.3.7
numpy==1.13.0
olefile==0.44
pandas==0.20.2
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
ptpython==0.39
ptyprocess==0.5.1
pyasn1==0.2.3
pycparser==2.17
pygal==2.3.1
Pygments==2.2.0
pymongo==3.4.0
pysolr==3.6.0
python-dateutil==2.6.0
pytz==2017.2
rcssmin==1.0.6
requests==2.18.1
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
